+++ src/widgets/visualizer/ui/chart/ui/bar/index.tsx
